Main Changes:
 - Updated the simulation time so that the underlying seconds variable is a double
 - Updated the generators so that they produce doubles

+++ call-action.h
@@ -21,7 +21,7 @@ namespace Project2
         
         virtual void Invoke() override
         {
-            uint32_t rndNum = generator->Next();
+            double rndNum = generator->Next();
 #ifdef TRACE
             stringstream ss;
